Add WIP plugin to the ops repository and add hold label
As part of https://github.com/gitpod-io/ops/issues/676 we discovered
that Tide tried to merge draft PRs in the ops repository.

By adding the WIP plugin the do-not-merge/work-in-progress label is
applied to DRAFT PRs. The Tide config for ops has been updated to
exlucde PRs with that labal.

This follows the setup for the other repositories.