Fix bug with missing PRs
The problem with the existing GraphQL query was that it used the commit
as a cursor but when doing so it doesn't include the commit itself.

WIth this PR we now use the object query to fetch the commit and then
ask for the history without using a cursor.

I also

- Updated the README to include instructions on how to establish a
  dev-loop for the service and instructions on how to deploy new
  versions.
- Added more logging to make it easier to debug what the service is
  doing - this was useful during development might help us debug cases
  in the future where it doesn't apply a label when we expect it to.